Refactoring to multiple src files
