+++ benchmark.py
@@ -1,151 +1,17 @@
 #!/usr/bin/env python3
 
-import argparse
 import concurrent.futures
-import datetime
-import enum
-import os
 import os.path
 import pprint
 import subprocess
 import sys
-import yaml
 
-from mlperf_logging import mllog
+from benchmark.cli import parse_arguments, validate_args
+from benchmark.config import *
+from benchmark.logging import setup_logging
+from benchmark.rules import validate_dlio_parameter
+from benchmark.utils import read_config_from_file, update_nested_dict, create_nested_dict
 
-import logging
-
-# Define the custom log levels
-STATUS = 25
-VERBOSE = 13
-VERBOSER = 12
-VERBOSEST = 11
-
-STREAM_LOG_LEVEL = logging.DEBUG
-
-COLOR_MAP = {
-    'normal': "\033[0m",
-    'white': "\033[37m",
-    "green": "\033[32m",
-    "yellow": "\033[33m",
-    "red": "\033[31m",
-    "magenta": "\033[35m",
-    "cyan": "\033[36m",
-    'intense_purple': '\033[1;95m',
-}
-
-# Create a custom logger
-logger = logging.getLogger('custom_logger')
-logger.setLevel(STREAM_LOG_LEVEL)
-
-# Define the custom log levels
-logging.addLevelName(VERBOSE, "VERBOSE")
-logging.addLevelName(VERBOSER, "VERBOSER")
-logging.addLevelName(VERBOSEST, "VERBOSEST")
-logging.addLevelName(STATUS, "STATUS")
-
-
-# Define the custom log level methods
-def verbose(self, message, *args, **kwargs):
-    if self.isEnabledFor(VERBOSE):
-        self._log(VERBOSE, message, args, **kwargs)
-
-
-def verboser(self, message, *args, **kwargs):
-    if self.isEnabledFor(VERBOSER):
-        self._log(VERBOSER, message, args, **kwargs)
-
-
-def verbosest(self, message, *args, **kwargs):
-    if self.isEnabledFor(VERBOSEST):
-        self._log(VERBOSEST, message, args, **kwargs)
-
-
-def status(self, message, *args, **kwargs):
-    if self.isEnabledFor(STATUS):
-        self._log(STATUS, message, args, **kwargs)
-
-
-# Create a custom formatter for the logger
-class ColoredFormatter(logging.Formatter):
-    def format(self, record):
-        color_map = {
-            logging.DEBUG: COLOR_MAP['white'],
-            logging.WARNING: COLOR_MAP['yellow'],
-            logging.ERROR: COLOR_MAP['red'],
-            logging.CRITICAL: COLOR_MAP['red'],
-            STATUS: COLOR_MAP['intense_purple'],
-        }
-        formatted_time = f"{datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')}"
-        color = color_map.get(record.levelno, COLOR_MAP['normal'])
-        return f"{color}{formatted_time}|{record.levelname}:{record.module}:{record.lineno}: " \
-               f"{record.getMessage()}{COLOR_MAP['normal']}"
-
-
-# Add the custom log level methods to the logger
-logging.Logger.verbose = verbose
-logging.Logger.verboser = verboser
-logging.Logger.verbosest = verbosest
-logging.Logger.status = status
-
-stream_handler = logging.StreamHandler()
-stream_handler.setFormatter(ColoredFormatter())
-stream_handler.setLevel(STREAM_LOG_LEVEL)
-logger.addHandler(stream_handler)
-
-
-# Define constants:
-COSMOFLOW = "cosmoflow"
-RESNET = "resnet50"
-UNET = "unet3d"
-MODELS = [COSMOFLOW, RESNET, UNET]
-
-H100 = "h100"
-A100 = "a100"
-ACCELERATORS = [H100, A100]
-
-OPEN = "open"
-CLOSED = "closed"
-CATEGORIES = [OPEN, CLOSED]
-
-LLAMA3_70B = 'llama3-70b'
-LLAMA3_405B = 'llama3-405b'
-LLM_1620B = 'llm-1620b'
-LLM_MODELS = [LLAMA3_70B, LLAMA3_405B, LLM_1620B]
-
-LLM_MODEL_PARALLELISMS = {
-    LLAMA3_70B: dict(tp=8, pp=4, min_ranks=4 * 8),
-    LLAMA3_405B: dict(tp=8, pp=16, min_ranks=8 * 16),
-    LLM_1620B: dict(tp=8, pp=64, min_ranks=64 * 8),
-}
-MIN_RANKS_STR = "; ".join(
-    [f'{key} = {value["min_ranks"]} accelerators' for key, value in LLM_MODEL_PARALLELISMS.items()])
-
-MPIRUN = "mpirun"
-MPIEXEC = "mpiexec"
-MPI_CMDS = [MPIRUN, MPIEXEC]
-
-STEPS_PER_EPOCH = 500
-MOST_MEMORY_MULTIPLIER = 5
-MAX_READ_THREADS_TRAINING = 32
-
-DEFAULT_HOSTS = "127.0.0.1"
-
-MPI_RUN_BIN = os.environ.get("MPI_RUN_BIN", MPIRUN)
-ALLOW_RUN_AS_ROOT = True
-
-class EXEC_TYPE(enum.Enum):
-    MPI = "mpi"
-    DOCKER = "docker"
-
-
-class PARAM_VALIDATION(enum.Enum):
-    CLOSED = "closed"
-    OPEN = "open"
-    INVALID = "invalid"
-
-
-DATETIME_STR = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
 
 # Capturing TODO Items:
 # Configure datasize to collect the memory information from the hosts instead of getting a number of hosts for the
@@ -158,146 +24,6 @@ DATETIME_STR = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
 # Determine method to use cgroups for memory limitation in the benchmark script.
 #
 # Add a log block at the start of datagen & run that output all the parms being used to be clear on what a run is.
-
-
-def parse_arguments():
-    # Many of the help messages are shared between the subparsers. This dictionary prevents rewriting the same messages
-    # in multiple places.
-    help_messages = dict(
-        model="Model to emulate. A specific model defines the sample size, sample container format, and data "
-              "rates for each supported accelerator.",
-        accelerator_type="Accelerator to simulate for the benchmark. A specific accelerator defines the data access "
-                         "sizes and rates for each supported workload",
-        num_accelerators_datasize="Simulated number of accelerators. In multi-host configurations the accelerators "
-                                  "will be initiated in a round-robin fashion to ensure equal distribution of "
-                                  "simulated accelerator processes",
-        num_accelerators_datagen="Number of parallel processes to use for dataset generation. Processes will be "
-                                 "initiated in a round-robin fashion across the configured client hosts",
-        num_client_hosts="Number of participating client hosts. Simulated accelerators will be initiated on these "
-                         "hosts in a round-robin fashion",
-        client_host_mem_GB="Memory available in the client where the benchmark is run. The dataset needs to be 5x the "
-                           "available memory for closed submissions.",
-        client_hosts="Comma separated IP addresses of the participating hosts (without spaces). "
-                     "eg: '192.168.1.1,192.168.1.2'",
-        category="Benchmark category to be submitted.",
-        results_dir="Directory where the benchmark results will be saved.",
-        params="Additional parameters to be passed to the benchmark. These will override the config file. For a closed "
-               "submission only a subset of params are supported. Multiple values allowed in the form: "
-               "--params key1=value1 key2=value2 key3=value3",
-        datasize="The datasize command calculates the number of samples needed for a given workload, accelerator type,"
-                 " number of accelerators, and client host memory.",
-        datagen="The datagen command generates a dataset for a given workload and number of parallel generation "
-                "processes.",
-        run_benchmark="Run the benchmark with the specified parameters.",
-        configview="View the final config based on the specified options.",
-        reportgen="Generate a report from the benchmark results.",
-
-        # Checkpointing help messages
-        checkpoint="The checkpoint command executes checkpoints in isolation as a write-only workload",
-        recovery="The recovery command executes a recovery of the most recently written checkpoint with randomly "
-                 "assigned reader to data mappings",
-        llm_model="The model & size to be emulated for checkpointing. The selection will dictate the TP, PP, & DP "
-                  "sizes as well as the size of the checkpoint",
-        num_checkpoint_accelerators=f"The number of accelerators to emulate for the checkpoint task. Each LLM Model "
-                                    f"can be executed as 8 accelerators or the minimum required to run the model: "
-                                    f"\n{MIN_RANKS_STR}"
-    )
-
-    parser = argparse.ArgumentParser(description="Script to launch the MLPerf Storage benchmark")
-    parser.add_argument("--allow-invalid-params", "-aip", action="store_true", help="Do not fail on invalid parameters.")
-    sub_programs = parser.add_subparsers(dest="program", required=True, help="Sub-programs")
-    sub_programs.required = True
-
-    # Training
-    training_parsers = sub_programs.add_parser("training", help="Training benchmark options")
-    training_parsers.add_argument("--data-dir", '-dd', type=str, help="Filesystem location for data")
-    training_parsers.add_argument('--results-dir', '-rd', type=str, required=True, help=help_messages['results_dir'])
-    training_subparsers = training_parsers.add_subparsers(dest="command", required=True, help="Sub-commands")
-    training_parsers.required = True
-
-    datasize = training_subparsers.add_parser("datasize", help=help_messages['datasize'])
-    datasize.add_argument('--model', '-m', choices=MODELS, required=True, help=help_messages['model'])
-    datasize.add_argument('--accelerator-type', '-g', choices=ACCELERATORS, required=True, help=help_messages['accelerator_type'])
-    datasize.add_argument('--num-accelerators', '-na', type=int, required=True, help=help_messages['num_accelerators_datasize'])
-    datasize.add_argument('--num-client-hosts', '-nc', type=int, required=True, help=help_messages['num_client_hosts'])
-    datasize.add_argument('--client-host-memory-in-gb', '-cm', type=int, required=True, help=help_messages['client_host_mem_GB'])
-
-    datagen = training_subparsers.add_parser("datagen", help=help_messages['datagen'])
-    datagen.add_argument('--hosts', '-s', type=str, default=DEFAULT_HOSTS, help=help_messages['client_hosts'])
-    datagen.add_argument('--category', '-c', choices=CATEGORIES, help=help_messages['category'])
-    datagen.add_argument('--model', '-m', choices=MODELS, required=True, help=help_messages['model'])
-    datagen.add_argument('--accelerator-type', '-a', choices=ACCELERATORS, required=True, help=help_messages['accelerator_type'])
-    datagen.add_argument('--num-accelerators', '-n', type=int, required=True, help=help_messages['num_accelerators_datagen'])
-    datagen.add_argument('--params', '-p', nargs="+", type=str, help=help_messages['params'])
-
-    run_benchmark = training_subparsers.add_parser("run", help=help_messages['run_benchmark'])
-    run_benchmark.add_argument('--hosts', '-s', type=str, default=DEFAULT_HOSTS, help=help_messages['client_hosts'])
-    run_benchmark.add_argument('--category', '-c', choices=CATEGORIES, help=help_messages['category'])
-    run_benchmark.add_argument('--model', '-m', choices=MODELS, required=True, help=help_messages['model'])
-    run_benchmark.add_argument('--accelerator-type', '-a', choices=ACCELERATORS, required=True, help=help_messages['accelerator_type'])
-    run_benchmark.add_argument('--num-accelerators', '-n', type=int, required=True, help=help_messages['num_accelerators_datasize'])
-    run_benchmark.add_argument('--params', '-p', nargs="+", type=str, help=help_messages['params'])
-
-    configview = training_subparsers.add_parser("configview", help=help_messages['configview'])
-    configview.add_argument('--model', '-m', choices=MODELS, help=help_messages['model'])
-    configview.add_argument('--accelerator-type', '-a', choices=ACCELERATORS, help=help_messages['accelerator_type'])
-    configview.add_argument('--params', '-p', nargs="+", type=str, help=help_messages['params'])
-
-    reportgen = training_subparsers.add_parser("reportgen", help=help_messages['reportgen'])
-    reportgen.add_argument('--results-dir', '-r', type=str, help=help_messages['results_dir'])
-
-    mpi_options = training_parsers.add_argument_group("MPI")
-    mpi_options.add_argument('--oversubscribe', action="store_true")
-    # mpi_options.add_argument('--allow-run-as-root', action="store_true")
-
-
-    # Checkpointing
-    checkpointing_parsers = sub_programs.add_parser("checkpointing", help="Checkpointing benchmark options")
-    checkpointing_subparsers = checkpointing_parsers.add_subparsers(dest="command", required=True, help="Sub-commands")
-    checkpointing_parsers.required = True
-
-    # Add specific checkpointing benchmark options here
-    checkpoint = checkpointing_subparsers.add_parser('checkpoint', help=help_messages['checkpoint'])
-    checkpoint.add_argument('--llm-model', '-lm', choices=LLM_MODELS, help=help_messages['llm_model'])
-    checkpoint.add_argument('--hosts', '-s', type=str, help=help_messages['client_hosts'])
-    checkpoint.add_argument('--num-accelerators', '-na', type=int, help=help_messages['num_checkpoint_accelerators'])
-
-    recovery = checkpointing_subparsers.add_parser('recovery', help=help_messages['recovery'])
-
-    # VectorDB Benchmark
-    vectordb_parsers = sub_programs.add_parser("vectordb", help="VectorDB benchmark options")
-    vectordb_parsers.add_argument('--hosts', '-s', type=str, help=help_messages['client_hosts'])
-
-    vectordb_subparsers = vectordb_parsers.add_subparsers(dest="command", required=True, help="Sub-commands")
-    vectordb_parsers.required = True
-
-    # Add specific VectorDB benchmark options here
-    throughput = vectordb_subparsers.add_parser('throughput', help=help_messages['vectordb_throughput'])
-    latency = vectordb_subparsers.add_parser('latency', help=help_messages['vectordb_latency'])
-
-    return parser.parse_args()
-
-
-def validate_args(args):
-    error_messages = []
-    # Add generic validations here. Workload specific validation is in the Benchmark classes
-
-    if error_messages:
-        for msg in error_messages:
-            print(msg)
-
-        sys.exit(1)
-
-
-def read_config_from_file(relative_path):
-    config_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), relative_path)
-    if not os.path.isfile(config_path):
-        raise FileNotFoundError(f"Configuration file not found: {config_path}")
-
-    with open(config_path, 'r') as f:
-        config = yaml.safe_load(f)
-
-    return config
 
 
 def generate_mpi_prefix_cmd(mpi_cmd, hosts, num_processes, oversubscribe, allow_run_as_root):
@@ -316,63 +42,6 @@ def generate_mpi_prefix_cmd(mpi_cmd, hosts, num_processes, oversubscribe, allow_
 
     return prefix
 
-
-def validate_dlio_parameter(model, param, value):
-    """
-    This function applies rules to the allowed changeable dlio parameters.
-    """
-    if model in MODELS:
-        # Allowed to change data_folder and number of files to train depending on memory requirements
-        if param.startswith('dataset'):
-            left, right = param.split('.')
-            if right in ('data_folder', 'num_files_train'):
-                # TODO: Add check of min num_files for given memory config
-                return PARAM_VALIDATION.CLOSED
-
-        # Allowed to set number of read threads
-        if param.startswith('reader'):
-            left, right = param.split('.')
-            if right == "read_threads":
-                if 0 < int(value) < MAX_READ_THREADS_TRAINING:
-                    return PARAM_VALIDATION.CLOSED
-
-    elif model in LLM_MODELS:
-        # TODO: Define params that can be modified in closed
-        pass
-
-    return PARAM_VALIDATION.INVALID
-
-
-def update_nested_dict(original_dict, update_dict):
-    updated_dict = {}
-    for key, value in original_dict.items():
-        if key in update_dict:
-            if isinstance(value, dict) and isinstance(update_dict[key], dict):
-                updated_dict[key] = update_nested_dict(value, update_dict[key])
-            else:
-                updated_dict[key] = update_dict[key]
-        else:
-            updated_dict[key] = value
-    for key, value in update_dict.items():
-        if key not in original_dict:
-            updated_dict[key] = value
-    return updated_dict
-
-
-def create_nested_dict(flat_dict, parent_dict=None, separator='.'):
-    if parent_dict is None:
-        parent_dict = {}
-
-    for key, value in flat_dict.items():
-        keys = key.split(separator)
-        current_dict = parent_dict
-        for i, k in enumerate(keys[:-1]):
-            if k not in current_dict:
-                current_dict[k] = {}
-            current_dict = current_dict[k]
-        current_dict[keys[-1]] = value
-
-    return parent_dict
 
 
 class ClusterInformation:
@@ -667,5 +336,5 @@ def main(args):
 if __name__ == "__main__":
     # Get the mllogger and args. Call main to run program
     cli_args = parse_arguments()
-    logger.setLevel(logging.DEBUG)
+    logger = setup_logging("MLPerfStorage")
     main(cli_args)
